fix: added checks for project existing in multiple project business logic functions

+++ app/service/project.py
@@ -56,10 +56,16 @@ async def add_todo_service(
     project_id: str, todo_request: AddTodoRequest, db: AsyncDatabase
 ) -> AddTodoResponse:
 
+    # Check if project exists
+    project_in_db_dict = await db_get_project(project_id, db)
+    if not project_in_db_dict:
+        raise HTTPException(
+            status_code=404, detail=f"Project does not exist: project_id={project_id}"
+        )
+
     # Assign to the first status_id in the project if status_id is not passed in
     if todo_request.status_id is None:
         # Get the project's current statuses
-        project_in_db_dict = await db_get_project(project_id, db)
         if project_in_db_dict["todo_statuses"]:
             todo_request.status_id = project_in_db_dict["todo_statuses"][0]["id"]
 
@@ -79,7 +85,7 @@ async def update_todo_service(
             status_code=404, detail=f"Project does not exist: project_id={project_id}"
         )
 
-    # Check if status_id exists in project's statuses
+    # Check if status_id exists in project
     status_ids = [str(status["id"]) for status in project_in_db_dict["todo_statuses"]]
     if update_todo_request.status_id not in status_ids:
         raise HTTPException(
@@ -98,6 +104,20 @@ async def delete_todo_service(
     db: AsyncDatabase,
 ) -> DeleteTodoResponse:
 
+    # Check if project exists
+    project_in_db_dict = await db_get_project(project_id, db)
+    if not project_in_db_dict:
+        raise HTTPException(
+            status_code=404, detail=f"Project does not exist: project_id={project_id}"
+        )
+
+    # Check if todo exists
+    if delete_todo_request.todo_id not in project_in_db_dict["todo_ids"]:
+        raise HTTPException(
+            status_code=404,
+            detail=f"Todo does not exist in project: todo_id={delete_todo_request.todo_id}, project_id={project_id}",
+        )
+
     await db_delete_todo(project_id, delete_todo_request.todo_id, db)
 
     return DeleteTodoResponse()
@@ -106,6 +126,13 @@ async def delete_todo_service(
 async def get_todo_items_service(
     project_id: str, db: AsyncDatabase
 ) -> GetTodoItemsResponse:
+
+    # Check if project exists
+    project_in_db_dict = await db_get_project(project_id, db)
+    if not project_in_db_dict:
+        raise HTTPException(
+            status_code=404, detail=f"Project does not exist: project_id={project_id}"
+        )
 
     todo_items_in_db_list = await db_get_todo_items(project_id, db)
 
@@ -148,6 +175,13 @@ async def add_todo_status_service(
     project_id: str, todo_status_request: AddTodoStatusRequest, db: AsyncDatabase
 ) -> AddTodoStatusResponse:
 
+    # Check if project exists
+    project_in_db_dict = await db_get_project(project_id, db)
+    if not project_in_db_dict:
+        raise HTTPException(
+            status_code=404, detail=f"Project does not exist: project_id={project_id}"
+        )
+
     await db_add_todo_status(project_id, todo_status_request.name, db)
 
     return AddTodoStatusResponse()
@@ -173,10 +207,15 @@ async def reorder_todo_statuses_service(
     db: AsyncDatabase,
 ) -> ReorderTodoStatusesResponse:
 
+    # TODO: Validate new_status_ids contains all existing status ids exactly once
     new_status_ids = reorder_todo_statuses_request.new_status_ids
 
     # Get the project's current statuses
     project_in_db_dict = await db_get_project(project_id, db)
+    if not project_in_db_dict:
+        raise HTTPException(
+            status_code=404, detail=f"Project does not exist: project_id={project_id}"
+        )
 
     new_statuses = sorted(
         project_in_db_dict["todo_statuses"],
